[CHANGE] mv some utilities to its right place
