+++ domain/GP_Prior.py
@@ -5,7 +5,7 @@ from adapters.gpytorch.means import LinearMean
 from gpytorch.means import ConstantMean
 from adapters.gpytorch.kernels import get_linear_kernel, get_squared_exponential_kernel, get_matern32_kernel, get_matern52_kernel, get_spectral_mixture_kernel, get_rff_kernel, get_polynomial_d2_kernel, get_polynomial_d3_kernel, get_polynomial_d4_kernel, get_piecewise_polynomial_kernel, get_base_kernels, wrap_scale_kernel, additive_structure_kernel
 from domain.env import KERNEL_TYPE
-from adapters.util import get_feature_names_from_rv_id, print_scores, get_err_dict
+from adapters.util import get_feature_names_from_rv_id, print_scores, get_err_dict, weighted_avg_and_std, is_positive_semi_definite
 import math
 from scipy.special import binom
 import jax.numpy as jnp
@@ -13,26 +13,6 @@ import numpy as np
 import time
 import copy
 from abc import abstractmethod
-
-def weighted_avg_and_std(values, weights, gamma=1):
-    """
-    Return the weighted average and standard deviation.
-
-    values, weights -- Numpy ndarrays with the same shape.
-    """
-    if gamma != 1:
-        weights = np.power(weights, gamma)
-    average = np.average(values, weights=weights)
-    variance = np.average((values - average) ** 2, weights=weights)
-    if variance <= 0:
-        sqr_var = 0.0
-    else:
-        sqr_var = math.sqrt(variance)
-    return average, sqr_var
-
-def is_positive_semi_definite(matrix):
-    return np.all(np.linalg.eigvals(matrix) >= 0)
-
 
 class Priors:
     def __init__(self, X, y, feature_names):
